agents/requirements.txt
pymupdf
cohere
qdrant-client
python-dotenv
langchain-qdrant
langchain-cohere
langchain_openai 
langgraph 
typing 
datetime
pandas
matplotlib
yfinance
uvicorn
fastapi